[remove-analysis-collection] Add getAnalysisList method to the analysis-service
