+++ src/windshaft/map-serializer/anonymous-map-serializer/analysis-serializer.js
@@ -1,14 +1,13 @@
 var _ = require('underscore');
-var LayerTypes = require('../../../geo/map/layer-types.js');
+var AnalysisService = require('../../../analysis/analysis-service');
 
 /**
  * Return a payload with the serialization of all the analyses in the 
  * layersCollection and the dataviewsCollection.
  */
 function serialize (layersCollection, dataviewsCollection) {
-  var layerAnalyses = _getAnalysesFromLayers(layersCollection);
-  var dataviewsAnalyses = _getAnalysesFromDataviews(dataviewsCollection);
-  return _generateUniqueAnalysisList(layerAnalyses.concat(dataviewsAnalyses));
+  var analysisList = AnalysisService.getAnalysisList(layersCollection, dataviewsCollection);
+  return _generateUniqueAnalysisList(analysisList);
 }
 
 /**
@@ -23,25 +22,6 @@ function _generateUniqueAnalysisList (analysesList) {
     }
     return list;
   }, []);
-}
-
-function _getAnalysesFromLayers (layersCollection) {
-  var layers = _getCartoDBAndTorqueLayers(layersCollection);
-  return layers.map(function (layer) {
-    return layer.getSource();
-  });
-}
-
-function _getAnalysesFromDataviews (dataviewsCollection) {
-  return dataviewsCollection.map(function (dataview) {
-    return dataview.getSource();
-  });
-}
-
-function _getCartoDBAndTorqueLayers (layersCollection) {
-  return layersCollection.filter(function (layer) {
-    return LayerTypes.isCartoDBLayer(layer) || LayerTypes.isTorqueLayer(layer);
-  });
 }
 
 /**
